pyproject.toml
[build-system]
# Build system requirements for setuptools editable install
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
# Project metadata
name = "detecterreur"
version = "0.1.0"
description = "A Python package to detect, correct, and provide educational feedback on errors in French written production for L2 learners."
readme = "README.md"
# FIX: Explicitly exclude Python 3.14 to allow uv sync to work automatically
requires-python = ">=3.10, <3.14"

# External dependencies required for the project
dependencies = [
    "codecarbon>=3.2.0",
    "matplotlib>=3.10.7",
    "prettytable>=3.17.0",
    "pygrammalecte>=1.5.0",
    "pyspellchecker>=0.8.4",
    "spacy>=3.8.11",
]

# Optional metadata
authors = [
    { name = "Salomon Metre", email = "salomon.metre8@etu.univ-lorraine.fr" },
    { name = "Fôtes Fataux Team" }
]
license = { text = "MIT" }
keywords = ["French", "L2 learning", "spellchecker", "error correction", "language learning"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Education",
    "Topic :: Text Processing :: Linguistic"
]

[tool.setuptools.packages.find]
# Look for packages inside the src/ directory
where = ["src"]